# Architecture_DZ_10/API/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cloud service for robot-vacuum cleaner",
    "description": "This is a study project considering server based on the OpenAPI 3.0 specification. \nYou can now help us improve the API whether it's by making changes to the definition itself or to the code.\nThat way, with time, we can improve the API in general, and expose some of the new features in OAS3.",
    "contact": {
      "email": "kardan97@yandex.ru"
    },
    "version": "0.0.2"
  },
  "servers": [
    {
      "url": "https://petstore3.swagger.io/api/v3"
    },
    {
      "url": "https://localhost:8080/api/v1"
    }
  ],
...
  "paths": {
...
    "/user": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Create user account",
        "description": "This create a new user account.",
        "operationId": "createUser",
        "requestBody": {
          "description": "Created user account",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/user/login": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Logs user into the system",
        "description": "",
        "operationId": "loginUser",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "description": "The user name for login",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "description": "The password for login in clear text",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "headers": {
              "X-Rate-Limit": {
                "description": "calls per hour allowed by the user",
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "X-Expires-After": {
                "description": "date in UTC when token expires",
                "schema": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            },
            "content": {
              "application/xml": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid username/password supplied"
          }
        }
      }
    },
...
    "/user/{userID}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get user by ID",
        "description": "Get user by ID",
        "operationId": "getUser",
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "description": "The user ID that needs to be returned",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid user ID supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Update user",
        "description": "This update info in user account.",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "description": "user id that need to be updated",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Update info in existing user account",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Delete user",
        "description": "This delete user account.",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "description": "user ID that needs to be deleted",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid user ID supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "country": {
            "$ref": "#/components/schemas/Country"
          },
          "city": {
            "type": "string"
          },
          "phone": {
            "type": "integer",
            "format": "int32",
            "example": 89991234567
          },
          "email": {
            "type": "string",
            "example": "john@email.com"
          },
          "password": {
            "type": "string"
          },
          "robotID1": {
            "$ref": "#/components/schemas/Robot"
          },
          "robotID2": {
            "$ref": "#/components/schemas/Robot"
          },
          "robotID3": {
            "$ref": "#/components/schemas/Robot"
          }
        },
        "xml": {
          "name": "user"
        }
      },
      "Country": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "cityID": {
            "$ref": "#/components/schemas/City"
          }
        },
        "xml": {
          "name": "country"
        }
      },
      "City": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        },
        "xml": {
          "name": "city"
        }
      },
      "Statistic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "timetableID": {
            "$ref": "#/components/schemas/Timetable"
          },
          "resultID": {
            "type": "string"
          }
        },
        "xml": {
          "name": "statistic"
        }
      },
      "Timetable": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "datetime": {
            "type": "string",
            "format": "date-time"
          },
          "roomID": {
            "$ref": "#/components/schemas/Room"
          },
          "resultID": {
            "$ref": "#/components/schemas/Result"
          }
        },
        "xml": {
          "name": "timetable"
        }
      },
      "Room": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "roomname": {
            "type": "string"
          },
          "roomsquare": {
            "type": "string"
          }
        },
        "xml": {
          "name": "room"
        }
      },
      "Result": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "Details": {
            "type": "string",
            "description": "What does this code mean"
          }
        },
        "xml": {
          "name": "result"
        }
      },
      "Model": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "firmwareID": {
            "$ref": "#/components/schemas/Firmware"
          }
        },
        "xml": {
          "name": "model"
        }
      },
      "Firmware": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "version": {
            "type": "string"
          },
          "URL": {
            "type": "string"
          }
        },
        "xml": {
          "name": "firmware"
        }
      },
      "Robot": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "serialNo": {
            "type": "string"
          },
          "model": {
            "$ref": "#/components/schemas/Model"
          },
          "firmware": {
            "type": "string"
          },
          "config": {
            "type": "array",
            "xml": {
              "wrapped": true
            },
            "items": {
              "type": "string",
              "xml": {
                "name": "config"
              }
            }
          },
          "diadnostic": {
            "type": "array",
            "xml": {
              "wrapped": true
            },
            "items": {
              "type": "integer",
              "format": "int32",
              "xml": {
                "name": "diadnostic"
              }
            }
          },
          "historyClean": {
            "$ref": "#/components/schemas/Statistic"
          },
          "status": {
            "type": "string",
            "description": "robot status in the store",
            "enum": [
              "ready",
              "charging",
              "working"
            ]
          }
        },
        "xml": {
          "name": "robot"
        }
      },
...
    },
...
  }
}